+++ .nuget/Changelog.md
@@ -1,3 +1,11 @@
 Element extensions
 
 - New GetParameter(ForgeTypeId) extension
+- New Cast<T>() extension
+- Removed CanBeNull attribute for ToElement<T> extension
+
+Host Extensions
+
+- New GetBottomFaces() extension
+- New GetTopFaces() extension
+- New GetSideFaces() extension